part addMasker
