+++ lib/pages/child/part_addMasker.dart
@@ -1,0 +1,38 @@
+import 'package:flutter/material.dart';
+
+class AddMasker extends StatelessWidget {
+  final String text;
+  final int fillColor;
+  final int textColor;
+  final double textSize;
+  final Function callback;
+
+  const AddMasker({
+    Key key,
+    this.fillColor,
+    this.text,
+    this.textColor = 0xFF000000,
+    this.textSize = 24,
+    this.callback,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      child: SizedBox(
+        width: 65,
+        height: 90,
+        child: FlatButton(
+          shape:
+              RoundedRectangleBorder(borderRadius: BorderRadius.circular(50)),
+          child: Text(text, style: Theme.of(context).textTheme.subtitle1),
+          onPressed: () {
+            callback(text);
+          },
+          color: fillColor != null ? Color(fillColor) : null,
+          textColor: Color(textColor),
+        ),
+      ),
+    );
+  }
+}
